feat(Game.kt): split Controller into Game and Controller classes
